tasks/views.py
from django.shortcuts import render
from rest_framework import viewsets
from .models import Task
from .serializers import TaskSerializer

# API ViewSet - Create, Read, Update, Delete
class TaskViewSet(viewsets.ModelViewSet):
    queryset = Task.objects.all().order_by('-created_at')
    serializer_class = TaskSerializer

# HTML page
def index(request):
    return render(request, 'index.html')
